refactor(incident): move INCIDENT_PAGE_FEATURE_FLAGS to shared utils location
